# [web] More efficient encoding of fallback font ranges