This PR changes the encoding of the NotoFont unicode ranges to a more compact form, to about 35% of the size of the previous encoding. This saves ~90KB from the size of every main.dart.js. Reducing the size of the ranges by ~3x reduces the total size of NotoFont data by a little over 2x.

Code size for counter demo:

|  |  Before | After  | Δ |
| --- | ---: | ---: | ---: |
| main.dart.js | 1583722 | 1490238 | -93534 (-5.91%) |
| brotli -9 | 426247 | 401639 | -24608 (-5.77%) |

The time spent parsing the ranges is reduced from 10.4ms to 4.1ms (median of 7 runs, quite variable, but clearly faster).

## Issues

- https://github.com/flutter/flutter/issues/131440

## Pre-launch Checklist

- [x] I read the [Contributor Guide] and followed the process outlined there for submitting PRs.
- [x] I read the [Tree Hygiene] wiki page, which explains my responsibilities.
- [x] I read and followed the [Flutter Style Guide] and the [C++, Objective-C, Java style guides].
- [x] I listed at least one issue that this PR fixes in the description above.
- [x] I added new tests to check the change I am making or feature I am adding, or Hixie said the PR is test-exempt. See [testing the engine] for instructions on writing and running engine tests.
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] I signed the [CLA].
- [ ] All existing and new tests are passing.

If you need help, consider asking for advice on the #hackers-new channel on [Discord].

<!-- Links -->
[Contributor Guide]: https://github.com/flutter/flutter/wiki/Tree-hygiene#overview
[Tree Hygiene]: https://github.com/flutter/flutter/wiki/Tree-hygiene
[Flutter Style Guide]: https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo
[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style
[testing the engine]: https://github.com/flutter/flutter/wiki/Testing-the-engine
[CLA]: https://cla.developers.google.com/
[flutter/tests]: https://github.com/flutter/tests
[breaking change policy]: https://github.com/flutter/flutter/wiki/Tree-hygiene#handling-breaking-changes
[Discord]: https://github.com/flutter/flutter/wiki/Chat